requirements.txt
python-telegram-bot==20.2
google-generativeai==0.3.0
PyYAML==6.0.1
python-dotenv==1.0.0
python-pptx==0.6.21
python-docx==0.8.11
aiohttp==3.10.0
